requirements.txt
Flask==1.1.1
Flask_SQLAlchemy==2.5.1
gunicorn==20.1.0
